requirements.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.3.0
python-decouple==3.8
